% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statamerge.R
\name{statamerge}
\alias{statamerge}
\title{Title}
\usage{
statamerge(
  master,
  using,
  mergetype = "1:1",
  merge_vars,
  suffix1 = ".x",
  suffix2 = ".y"
)
}
\arguments{
\item{master}{A dataframe containing the primary data}

\item{using}{A dataframe containing the data to merge to the primary data}

\item{mergetype}{A string that denotes the type of merge to perform (1:1, m:1, 1:m)}

\item{merge_vars}{A vector of the variables to merge the data on}

\item{suffix1}{A string denoting an alternate intermediate suffix for the join (.x)}

\item{suffix2}{A string denoting an alternate intermediate suffix for the join (.y)}
}
\value{
A merged dataframe
}
\description{
Title
}
\examples{

statamerge(dplyr::band_instruments, dplyr::band_members, mergetype = "1:1", merge_vars = "name")

}
